Group all config source commands into a subcommand
